1. implementasi api get menu 2. develop generate menu by bread 3. create default menu 4. create configuration page

+++ src/Controllers/BadasoBaseController.php
@@ -14,6 +14,11 @@ class BadasoBaseController extends Controller
         try {
             $slug = $this->getSlug($request);
             $data_type = $this->getDataType($slug);
+
+            if (!$this->isAuthorize('browse', $data_type)) {
+                return ApiResponse::forbidden();
+            }
+
             $data = $this->getDataList($slug, $request->all());
 
             return ApiResponse::entity($data_type, $data);
@@ -30,6 +35,9 @@ class BadasoBaseController extends Controller
             ]);
             $slug = $this->getSlug($request);
             $data_type = $this->getDataType($slug);
+            if (!$this->isAuthorize('read', $data_type)) {
+                return ApiResponse::forbidden();
+            }
             $data = $this->getDataDetail($slug, $request->id);
 
             return ApiResponse::entity($data_type, $data);
@@ -50,6 +58,9 @@ class BadasoBaseController extends Controller
             ]);
             $slug = $this->getSlug($request);
             $data_type = $this->getDataType($slug);
+            if (!$this->isAuthorize('edit', $data_type)) {
+                return ApiResponse::forbidden();
+            }
             $data = $this->createDataFromRaw($request->input('data') ?? [], $data_type);
             $this->validateData($data, $data_type);
             $updated_data = $this->updateData($data, $data_type);
@@ -76,6 +87,9 @@ class BadasoBaseController extends Controller
             ]);
             $slug = $this->getSlug($request);
             $data_type = $this->getDataType($slug);
+            if (!$this->isAuthorize('add', $data_type)) {
+                return ApiResponse::forbidden();
+            }
             $data = $this->createDataFromRaw($request->input('data') ?? [], $data_type);
             $this->validateData($data, $data_type);
             $stored_data = $this->insertData($data, $data_type);
@@ -102,6 +116,9 @@ class BadasoBaseController extends Controller
             ]);
             $slug = $this->getSlug($request);
             $data_type = $this->getDataType($slug);
+            if (!$this->isAuthorize('delete', $data_type)) {
+                return ApiResponse::forbidden();
+            }
             $data = $this->createDataFromRaw($request->input('data') ?? [], $data_type);
             $this->deleteData($data, $data_type);
 
@@ -127,6 +144,9 @@ class BadasoBaseController extends Controller
             ]);
             $slug = $this->getSlug($request);
             $data_type = $this->getDataType($slug);
+            if (!$this->isAuthorize('delete', $data_type)) {
+                return ApiResponse::forbidden();
+            }
             $data = $this->createDataFromRaw($request->input('data') ?? [], $data_type);
             $ids = $data['ids'];
             $id_list = explode(',', $ids);
